pyproject.toml
[project]
name = "python-robot"
version = "0.1.0"
description = "robot project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "websockets>=11.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "aliyun-python-sdk-core>=2.13.0",
    "dashscope>=1.14.0",
    "langchain<1.0.0",
    "langchain-community<1.0.0",
    "langchain-core<1.0.0",
    "chardet>=5.2.0",
    "langchain-chroma>=0.2.6",
    "chromadb>=1.4.1",
    "pypdf>=6.6.2",
    "unstructured>=0.18.31",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "python-multipart>=0.0.6",
    "mcp>=1.26.0",
    "git-filter-repo>=2.47.0",
]
